+++ src/forms/rating/adminedit.rs
@@ -15,15 +15,15 @@ pub struct AdminEditRating {
 impl AdminEditRating {
     pub fn update(self, rating: &mut models::Rating) {
         if let Some(comment) = self.comment {
-            rating.comment.insert(comment);
+            rating.comment = Some(comment);
         }
 
         if let Some(rate) = self.rate {
-            rating.rate.insert(rate);
+            rating.rate = Some(rate);
         }
 
         if let Some(hidden) = self.hidden {
-            rating.hidden.insert(hidden);
+            rating.hidden = Some(hidden);
         }
     }
 }